# Conflicts:
#	stp-domain/src/main/java/com/sytoss/domain/bom/personalexam/PersonalExam.java
#	stp-domain/src/main/java/com/sytoss/domain/bom/users/AbstractUser.java
#	stp-domain/src/main/java/com/sytoss/domain/bom/users/Group.java
#	stp-ms-test-producer/src/test/java/com/sytoss/producer/controllers/PersonalExamControllerTest.java

+++ stp-ms-lessons/src/test/java/com/sytoss/lessons/bdd/given/GroupGiven.java
@@ -54,7 +54,7 @@ public class GroupGiven extends LessonsIntegrationTest {
         }
 
         if (getTestExecutionContext().getIdMapping().get(disciplineId) != null) {
-            DisciplineDTO disciplineDTO = getDisciplineConnector().getById(getTestExecutionContext().getIdMapping().get(disciplineId));
+            DisciplineDTO disciplineDTO = getDisciplineConnector().getById((Long)getTestExecutionContext().getIdMapping().get(disciplineId));
             getDisciplineConnector().delete(disciplineDTO);
         }
